Makes conf file configurable at cluster start time
