+++ conf/kvs-base.yml
@@ -1,0 +1,11 @@
+ebs: ./
+threads:
+  memory: 4
+  ebs: 4
+  routing: 4
+  benchmark: 4
+replication:
+  memory: 1
+  ebs: 0
+  minimum: 1
+  local: 1
