+++ k8s/add_nodes.py
@@ -21,7 +21,7 @@ from util import *
 
 ec2_client = boto3.client('ec2')
 
-def add_nodes(client, kinds, counts, mon_ips, route_ips=[], node_ips=[]):
+def add_nodes(client, cfile, kinds, counts, mon_ips, route_ips=[], node_ips=[]):
     if node_ips:
         assert len(kinds) == len(counts) == len(node_ips), 'Must have same \
                 number of kinds and counts and node_ips.'
@@ -78,6 +78,7 @@ def add_nodes(client, kinds, counts, mon_ips, route_ips=[], node_ips=[]):
                 {'podid': podid}}})
             index += 1
 
+        created_pods = []
         print('Creating %d %s pod(s)...' % (counts[i], kind))
         for j in range(max_id + 1, max_id + counts[i] + 1):
             filename = 'yaml/pods/%s-pod.yml' % (kind)
@@ -113,4 +114,12 @@ def add_nodes(client, kinds, counts, mon_ips, route_ips=[], node_ips=[]):
 
             client.create_namespaced_pod(namespace=NAMESPACE,
                     body=pod_spec)
+
+            # wait until all pods of this kind are running
+            ips = get_pod_ips(client, 'role='+kind)
+            while len(ips) < counts[i]:
+                ips = get_pod_ips(client, 'role='+kind)
+
+            for pod in created_pods:
+                copy_file_to_pod(client, cfile, pod, '/fluent/conf/kvs-config.yml')
