+++ k8s/start.sh
@@ -41,19 +41,9 @@ else
   PUBLIC_IP=$PRIVATE_IP
 fi
 
-echo -e "threads:" > conf/kvs-config.yml
-echo -e "    memory: 4" >> conf/kvs-config.yml
-echo -e "    ebs: 4" >> conf/kvs-config.yml
-echo -e "    benchmark: 4" >> conf/kvs-config.yml
-echo -e "    routing: 4" >> conf/kvs-config.yml
-
-echo -e "replication:" >> conf/kvs-config.yml
-echo -e "    memory: 1" >> conf/kvs-config.yml
-echo -e "    ebs: 0" >> conf/kvs-config.yml
-echo -e "    minimum: 1" >> conf/kvs-config.yml
-echo -e "    local: 1" >> conf/kvs-config.yml
-
-echo -e "ebs: /ebs" >> conf/kvs-config.yml
+while [[ ! -f "conf/kvs-config.yml" ]]; do
+  continue
+done
 
 if [ "$1" = "mn" ]; then
   echo -e "monitoring:" >> conf/kvs-config.yml
